feat: product entity
