+++ src/main/java/com/daangn/daangn/domain/Product/Category.java
@@ -1,0 +1,24 @@
+package com.daangn.daangn.domain.Product;
+
+public enum Category {
+	DIGITALDEVICE,
+	HOMEAPPLIANCE,
+	FURNITURE,
+	KITCHEN,
+	KIDS,
+	KIDSBOOK,
+	WCLOTHES,
+	WSTUFF,
+	MFASHION,
+	BEAUTY,
+	SPORTS,
+	HOBBY,
+	CAR,
+	BOOK,
+	TICKET,
+	FOODS,
+	PET,
+	PLANT,
+	ETC
+
+}
