+++ src/main/java/com/daangn/daangn/domain/Product/ProductStatus.java
@@ -1,0 +1,5 @@
+package com.daangn.daangn.domain.Product;
+
+public enum ProductStatus {
+	ONSALE, RESERVED, OUTOFSTOCK, HIDDEN
+}
