--- src/main/java/com/daangn/daangn/dto/Category.java
@@ -1,15 +1,0 @@
-package com.daangn.daangn.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@AllArgsConstructor
-@Getter
-@Setter
-public class Category {
-
-	private Integer id;
-	private String name;
-
-}
