--- src/main/java/com/daangn/daangn/dto/ProductStatus.java
@@ -1,15 +1,0 @@
-package com.daangn.daangn.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@AllArgsConstructor
-@Getter
-@Setter
-public class ProductStatus {
-
-	private Integer id;
-	private String status;
-
-}
